+++ tests/read_json_file_test.go
@@ -35,8 +35,8 @@ import (
 	"testing"
 	"time"
 
-	integrations "github.com/ArrowArc/ArrowArc/internal/integrations/filesystem"
 	"github.com/apache/arrow/go/v17/arrow"
+	integrations "github.com/arrowarc/arrowarc/internal/integrations/filesystem"
 	"github.com/stretchr/testify/assert"
 )
