+++ tests/read_weather_api_test.go
@@ -34,9 +34,9 @@ import (
 	"testing"
 	"time"
 
-	weather "github.com/ArrowArc/ArrowArc/internal/integrations/api/weather"
-	"github.com/ArrowArc/ArrowArc/pkg/common/config"
-	helper "github.com/ArrowArc/ArrowArc/pkg/common/utils"
+	config "github.com/ArrowArc/ArrowArc/pkg/common/config"
+	weather "github.com/arrowarc/arrowarc/internal/integrations/api/weather"
+	helper "github.com/arrowarc/arrowarc/pkg/common/utils"
 	"github.com/stretchr/testify/assert"
 )
