+++ tests/write_json_file_test.go
@@ -35,10 +35,10 @@ import (
 	"testing"
 	"time"
 
-	integrations "github.com/ArrowArc/ArrowArc/internal/integrations/filesystem"
 	"github.com/apache/arrow/go/v17/arrow"
 	"github.com/apache/arrow/go/v17/arrow/array"
 	"github.com/apache/arrow/go/v17/arrow/memory"
+	integrations "github.com/arrowarc/arrowarc/internal/integrations/filesystem"
 	"github.com/stretchr/testify/assert"
 )
